BCM2835-V4L2: Optimisation to not remove padding where possible
Interleaved RGB and single plane YUV formats can be delivered
by the GPU without the secondary step of removing padding,
as the bytesperline field can be completed appropriately.

Planar YUV has to remain with the GPU removing padding when
necessary, as there is no way to report that there is
padding between the planes (ie on the height).
The multi-planar formats are NOT applicable, as there is no
easy way to make them contiguous in memory (ie one large
allocation that gets broken up). The whole task is passed
across to videobuf2 which has no notion of that requirement.