staging: bcm2835-camera: Remove bulk_mutex as it is not required
There is no requirement to serialise bulk transfers as that is
all done in VCHI, and if a second MMAL_MSG_TYPE_BUFFER_TO_HOST
happened before the VCHI_CALLBACK_BULK_RECEIVED, then the
service_callback thread is deadlocked.

Remove the bulk_mutex so that multiple receives can be
scheduled at a time.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>