staging: vc04_services: Add V4L2 M2M device for codecs.
Adds a V4L2 memory-to-memory device that wraps the video
encode and decode components on the VPU. This reuses a fair
chunk of the existing V4L2 camera driver MMAL VCHIQ interface.

Basic video encode is working. Controls need to be added.
Decode needs to sort out resolution changes and setting
selection regions.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.org>